Theme the docs to look more like webkit.org
MKDocs Material theme does not quite fit the theme of the WebKit project.
Let's make it look similar to the WebKit.org theme by porting over the CSS styles
used for the website.

We end up overriding a lot of the CSS which is not ideal, and
we should come up with a better solution in the future.