Improve styling to match webkit.org, and not rewrite the world
This commit build's on Martin Donath's work with not trying to override all
the variables and using a lot of the built-ins.

I am only doing minor restyling of the text to closely match webkit.org's style.
The rest of the styles should remain untouched. Colors are handled by over-riding the
MkDocs Material colors with Webkit's color scheme.