Rename ImportWPTTests to WPTTests
We should ideally have a single page that contains information about our
integration with WPT. Currently the page holds information for importing
WPT tests, but we should also add steps to exporting tests. Renaming
the title of the page to "Web Platform Tests Integration," makes it clear
that this is the location for that type of information.