Add prominent https://webkit.org/ link to top page
A developer/reader coming in to the https://docs.webkit.org/ site would
likely find it pretty useful to also browse directly to the top of the
https://webkit.org/ site and explore all the other information about WebKit
which that site provides.

But the existing https://docs.webkit.org/ site doesn’t include any link
directly to the top of the https://webkit.org/ site.

Among the specific things linked to on the https://webkit.org/ site,
developers are likely to find the Feature Status tab in the top menu
to be useful — since it includes links to the Standards Positions and
CSS Features pages, which docs.webkit.org doesn’t link to (and
probably doesn’t need to).

See also https://commits.webkit.org/267977@main